% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFragConfigR.R
\name{addMissingInChIKeys.CSV}
\alias{addMissingInChIKeys.CSV}
\title{Fill in Missing InChIKeys in CSV files}
\usage{
addMissingInChIKeys.CSV(csv_file, inchikey_col_number, smiles_col_number, babel_dir, write=TRUE, csv_file_out="")
}
\arguments{
\item{csv_file}{Full path and file name to CSV file to fill in InChIKey gaps}

\item{inchikey_col_number}{Column number of the column containing InChIKeys to check/fill}

\item{smiles_col_number}{Column number of the SMILES code needed to calculate missing InChIKeys}

\item{babel_dir}{Path to the directory containing \code{obabel.exe}, e.g. \code{"C:/Program Files (x86)/OpenBabel-2.3.2"}}

\item{write}{If \code{TRUE}, writes either to \code{csv_file_out} (of defined) or overwrites \code{csv_file} and
returns the file name. If \code{FALSE}, returns a list of InChIKeys and does not write to file.}

\item{csv_file_out}{Full path and file name of output CSV, define to avoid overwriting the original file.}
}
\value{
Returns path and file name to CSV file with fixed entries (if necessary) or list of InChIKeys.
}
\description{
This is a small wrapper function to patch CSV files that are missing some
InChIKey entries, using Cactus then Open Babel to fill gaps from SMILES.
}
\examples{
testCSV <- system.file("extdata","CompToxBatchSearch_MetFrag_MSready_C10H14N2_wSelectMetaData.csv",package="ReSOLUTION")
babel_dir <- "C:/Program Files (x86)/OpenBabel-2.3.2"
InChIKeys <- addMissingInChIKeys.CSV(testCSV,13,5,babel_dir,write=FALSE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{getInChIKey.obabel}}, \code{\link{getCactus}}, \code{\link{InChIKey_test}}
}

