% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EICandMSMS.R
\name{addMsMs.oldRMB}
\alias{addMsMs.oldRMB}
\title{Add MS/MS Scans to Plot - Backwards Compatible to Older RMassBank versions.}
\usage{
addMsMs.oldRMB(mzML_filePath, mz, rt, rt_window=NULL, maxCount=10, colour="black", isLog = FALSE)
}
\arguments{
\item{mzML_filePath}{Path to the mzML (or mzXML) file to extract the EICs.}

\item{mz}{mz value for MS/MS extraction.}

\item{rt}{rt value for MS/MS extraction (seconds).}

\item{rt_window}{RT window for extraction, as in \code{plotEICs}. If \code{NULL}, values
are extracted over whole run.}

\item{maxCount}{Maximum number of scans to extract, for \code{findMsMsHR.mass}.}

\item{colour}{Colour of the lines to plot. Default \code{black}; ideally same colour as EIC.}

\item{isLog}{Set whether log values should be plotted. Default \code{FALSE} - the setting
\code{TRUE} is completely untested (as EIC log plotting also not working)!}
}
\value{
Returns lines for the plot in the active device. If this returns errors, try
\code{\link{addMsMs}}. If this returns warnings, it is likely no MS/MS scans are present
matching the given limits (the warnings should indicate if this is the case).
}
\description{
This adds MS/MS Scans to EIC plots (or others) using \code{findMsMsHR.mass} from
\code{RMassBank}, compatible with pre 2.0.0 versions.
}
\examples{
library(RMassBank)
library(RMassBankData)
RmbDefaultSettings()
mzML_files <- list.files(system.file("spectra", package="RMassBankData"), ".mzML", full.names = TRUE)
mzML_file <- mzML_files[1]
cmpd_id <- "2818"
loadList(system.file("list/NarcoticsDataset.csv",package="RMassBankData"))
smiles <- findSmiles(cmpd_id)
mz <- as.numeric(findMz(cmpd_id, mode="pH")[3])
rt <- as.numeric(findRt(cmpd_id))
plotEICs(mzML_file, mz, rt, rt_in_sec=FALSE, smiles = smiles)
plotEICs(mzML_file, mz, rt, rt_in_sec=FALSE, smiles = smiles, addMSMS=TRUE, newRMB=FALSE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>, Michael Stravs <michael.stravs@eawag.ch>
}
\seealso{
\code{\link{findMsMsHR.mass}}, \code{\link{plotEICs}}, \code{\link{addMsMs}}.
}

