% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSMS.R
\name{addStructToPlot}
\alias{addStructToPlot}
\title{Add Structure to Plot (outdated)}
\usage{
addStructToPlot(smiles, kekulise = TRUE, coords = c(0, 1000, 100, 2500))
}
\arguments{
\item{smiles}{SMILES of structure to add to plot}

\item{kekulise}{Control aromaticity detection. Default \code{TRUE} is best with
recent \code{rcdk} versions.}

\item{coords}{The coordinates (xmin, ymin, xmax, ymax) for the structure.}
}
\value{
Adds the structure to the plot in the active device
}
\description{
This is an outdated function to add a structure to a plot,
kept for backwards compatability. Use \code{\link{renderSMILES.rcdk}} instead.
}
\examples{
ms1_mz <- c(210.1125, 211.1154, 212.1187)
ms1_int <- c(420889280, 52790856, 2265951.5)
ms1_labels <- c("[M+H]+", "13C-M+1", "13C-M+2")
plotMS(ms1_mz, ms1_int, main="test MS1 spec",labels=ms1_labels, formatFormula=FALSE,
absInt=FALSE)
addStructToPlot("O=C(Oc1ccccc1OC(C)C)NC", kekulise=TRUE, coords=c(205,1000,209,2000))
addStructToPlot("O=C(Oc1ccccc1OC(C)C)NC", kekulise=FALSE, coords=c(205,1000,209,2000))

}
\seealso{
\code{\link{renderSMILES.rcdk}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
