% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSMS.R
\name{chemistry2expression}
\alias{chemistry2expression}
\title{Convert Molecular Formulas to Regular Expression for Labels}
\usage{
chemistry2expression(formulas)
}
\arguments{
\item{formulas}{Vector containing formulas (as strings) to convert}
}
\value{
Expression for plotting labels.
}
\description{
Creates subscripts for numerals in molecular formulas for plotting
}
\examples{
labels_1 <- c("C2H4NO", "C6H7O2", "C8H10NO3", "C11H16NO3")
chemistry2expression(labels_1)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu> and Steffen Neumann <sneumann@ipb-halle.de>
}
\seealso{
\code{\link{plotSpectra}}, \code{\link{trimAnnotation}}, \code{\link{text}}.
}

