% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_resolution.R
\name{cluster_resolution}
\alias{cluster_resolution}
\title{cluster_resloution}
\usage{
cluster_resolution(graph, t = 1, directed = FALSE)
}
\arguments{
\item{graph}{An igraph network or a data frame of three columns: source, target, and weights.}

\item{t}{The time-scale parameter of the process which uncovers community structures at different resolutions.}

\item{directed}{Logical. TRUE if the network is directed. Ignored if graph is an igraph object.}
}
\value{
Table with information about community which has been found for each node.
}
\description{
Algorithm which finds communities using stability as an objective function to be optimised
                 in order to find the best partition  of network.
}

