% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_resolution_RandomOrderFULL.R
\name{cluster_resolution_RandomOrderFULL}
\alias{cluster_resolution_RandomOrderFULL}
\title{cluster_resolution_RandomOrderFULL}
\usage{
cluster_resolution_RandomOrderFULL(graph, t = 1, directed = FALSE,
  rep = 10)
}
\arguments{
\item{graph}{An igraph network or a data frame of three columns: source, target, and weights.}

\item{t}{The time-scale parameter of the process which uncovers community structures at different resolutions.}

\item{directed}{Logical. TRUE if the network is directed. Ignored if graph is an igraph object.}

\item{rep}{You can choose the number of repetitions. The default is 10.}
}
\value{
return list where we have (1) table with each result, (2) modularity for each outcome, (3) the best clustering,
(4) the the highest value of modularity which the best clustering has.
}
\description{
cluster_resolution_RandomOrderFULL is enlargement of function cluster_resolution in the case where outcome based on random
order of nodes. Function repeats rep times algorithm and return all the obtained results (see the return section).
}
\examples{
library(igraph)
g <- nexus.get("miserables")
cluster_resolution_RandomOrderFULL(g,directed=FALSE,t=1)
cluster_resolution_RandomOrderFULL(g,directed=FALSE,t=1,rep=20)
}

