% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFragConfigR.R
\name{CompToxCSVtoLocalCSVterms}
\alias{CompToxCSVtoLocalCSVterms}
\title{Prepare CompTox Dashboard MetFrag Export CSV files for MetFragCL}
\usage{
CompToxCSVtoLocalCSVterms(csv_file, start_index=15)
}
\arguments{
\item{csv_file}{Full path and file name to the Dashboard Export file to process}

\item{start_index}{The column number where the metadata columns start (default \code{15} is appropriate
for CompTox Dashboard MetFragBeta Export file format released March 2018)}
}
\value{
Returns a list containing the CSV file name, a list of scoring terms to add to
the MetFrag config file and a corresponding score weights entry.
}
\description{
This function prepares CompTox CSV Export files for use in MetFragCL. It
extracts numeric metadata field names for inclusion as scoring terms.
}
\examples{
CompToxCSV <- system.file("extdata","CompToxBatchSearch_MetFrag_MSready_C10H14N2_wSelectMetaData.csv",package="ReSOLUTION")
LocalCSVterms <- CompToxCSVtoLocalCSVterms(CompToxCSV)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu> in partnership with Christoph Ruttkies (MetFragCL),
Antony J. Williams and team (CompTox Dashboard)
}
\seealso{
\code{\link{MetFragConfig}}, \code{\link{runMetFrag}}
}

