% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFragConfigR.R
\name{CompToxSDFtoLocalSDFterms}
\alias{CompToxSDFtoLocalSDFterms}
\title{Prepare CompTox Dashboard MetFrag Export SDF files for MetFragCL}
\usage{
CompToxSDFtoLocalSDFterms(SDF_file)
}
\arguments{
\item{SDF_file}{Full path and file name to the Dashboard Export SDF file}
}
\value{
Returns a list containing the SDF file name, a list of scoring terms to add to
the MetFrag config file and a corresponding score weights entry.
}
\description{
This function prepares CompTox SDF MetFrag Export files for use in MetFragCL. It reads the
SDF and extracts numeric metadata field names for inclusion as scoring terms.
}
\examples{
CompToxSDF <- system.file("extdata","CompToxBatchSearch_MetFrag_MSready_C10H14N2.sdf",package="ReSOLUTION")
LocalSDFterms <- CompToxSDFtoLocalSDFterms(CompToxSDF)


}
\seealso{
\code{\link{MetFragConfig}}, \code{\link{runMetFrag}}, \code{\link{process.sdf.file}}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu> in partnership with Christoph Ruttkies (MetFragCL),
Antony J. Williams and team (CompTox Dashboard)
}
