% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_resolution.R
\name{Delta}
\alias{Delta}
\title{Delta}
\usage{
Delta(A, Adj, Com1, Com2)
}
\arguments{
\item{A}{Adjacency matrix}

\item{Adj}{Matrix calxulated from pattern e^(t(B-I))k_j}

\item{Com1}{node which is considered to change the community}

\item{Com2}{neighbors of node Com2}
}
\description{
Computes value Delta(R_NL) for each neighbor at the same time and returned vector delta with these values.

Delta(R_NL) evaluates the change of stability (R_NL) by removing Com1 from its community and
then by moving it into a neighbouring community. The node Com1 is then placed in the community for which
this gain is maximum, but only if this gain is positive.
}

