% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{Erase}
\alias{Erase}
\title{Erase}
\usage{
Erase(A, OldCom, NewCom)
}
\arguments{
\item{A}{Matrix which will be modified.}

\item{OldCom}{the community which the node have been.}

\item{NewCom}{the community which the node moved to.}
}
\description{
Function for the modification of the f.i. adjacency matrix resulting from merging communities OldCom and NewCom,
the row and column corresponding to OldCom are erased.
}

