% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{Exp_matrix}
\alias{Exp_matrix}
\title{Exp_matrix}
\usage{
Exp_matrix(A, p)
}
\arguments{
\item{A}{Adjacency matrix}

\item{p}{resolution parameter t}
}
\description{
Function for creating the matrix e^(t(B-I))k_j which is needed to computed stability.
}

