% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFragInterpreteR.R
\name{ExplFormulas2AnnoMZ}
\alias{ExplFormulas2AnnoMZ}
\title{Convert FormulasOfExplPeaks MetFrag Entry into MS/MS with mz and formulas}
\usage{
ExplFormulas2AnnoMZ(FormulasOfExplPeaks)
}
\arguments{
\item{FormulasOfExplPeaks}{A single \code{FormulasOfExplPeaks} entry from the MetFrag results file}
}
\value{
Returns a data frame with mz and formula fields (latter labeled \code{anno}).
}
\description{
This extracts the Explained MS/MS peak formulas from MetFrag results files
}
\examples{
FormulasOfExplPeaks <- "62.9638:[HO2P]-H-;78.959:[HO3P]-H-;153.0325:[C4H10O4P]-"
ExplFormulas2AnnoMZ(FormulasOfExplPeaks)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{

}

