% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFragInterpreteR.R
\name{ExplPeaks2MSMS}
\alias{ExplPeaks2MSMS}
\title{Convert ExplPeaks MetFrag Entry into MS/MS with mz and int}
\usage{
ExplPeaks2MSMS(ExplPeaks)
}
\arguments{
\item{ExplPeaks}{A single \code{ExplPeaks} entry from the MetFrag results file}
}
\value{
Returns a data frame with mz and int fields
}
\description{
This extracts the Explained MS/MS peaks from MetFrag results files
}
\examples{
ExplPeaks <- "63.9622_516.7;77.9655_999.0;82.0298_373.2"
ExplPeakList <- ExplPeaks2MSMS(ExplPeaks)

}
\seealso{

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
