% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_functions.R
\name{extractMS1.RMBarchive}
\alias{extractMS1.RMBarchive}
\title{Extract Raw MS1 Scan from RMassBank Archive}
\usage{
extractMS1.RMBarchive(RMBarchive, MS1_dir)
}
\arguments{
\item{RMBarchive}{Archive file to process}

\item{MS1_dir}{Directory in which to save the MS1 files}
}
\value{
Returns nothing; files are saved externally.
}
\description{
This extracts the "raw" MS1 data from a \code{RMassBank} archive file
and saves the corresponding MS1 files in the given directory.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{msmsWorkflow}}
}

