% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_functions.R
\name{extractMS1and2.RMBarchive}
\alias{extractMS1and2.RMBarchive}
\title{Extract MS1 and MS2 Peaklists from RMassBank Archive}
\usage{
extractMS1and2.RMBarchive(RMBarchive, MS1_dir, MS2_dir)
}
\arguments{
\item{RMBarchive}{Archive file to process}

\item{MS1_dir}{Directory in which to save the MS1 files}

\item{MS2_dir}{Directory in which to save the MS2 files}
}
\value{
Returns nothing; files are saved externally.
}
\description{
This extracts the "raw" MS1 data and the processed MS2 data
from a \code{RMassBank} archive file
and saves the corresponding files in the given directories. In contrast
to \code{\link{extractMS1.RMBarchive}}, this only saves data where MS2 were found.
}
\details{
Note that this function could be extended to improve the MS2 export.
The current state is very basic.
}
\seealso{
\code{\link{msmsWorkflow}}, \code{\link{extractMS1.RMBarchive}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
