% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassBankRecordeR.R
\name{fixMzIntColNames}
\alias{fixMzIntColNames}
\title{Fix MassBank Column Headers to be more R-friendly}
\usage{
fixMzIntColNames(ColNames)
}
\arguments{
\item{ColNames}{Column names to edit}
}
\value{
Returns edited column names.
}
\description{
This function renames column names internally in \code{\link{getMBPeaksAnnos}},
and \code{\link{getMBRecordPeaks}} to make them more R-friendly.
}
\details{
This function changes \code{m/z} to \code{mz}, replaces a decimal place with nothing
and changes \code{int} to \code{Int}.
}
\examples{
MBrecord <- system.file("extdata","EA020161_Diclofenac.txt",package="ReSOLUTION")
MBcolNames <- colnames(getMBRecordPeaks(MBrecord))
fixMzIntColNames(MBcolNames)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{getMBPeaksAnnos}}, \code{\link{getMBRecordPeaks}}.
}

