% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassBankRecordeR.R
\name{getMBPeaksAnnos}
\alias{getMBPeaksAnnos}
\title{Retrieve a Merged Peak List with Annotations from a MassBank Record}
\usage{
getMBPeaksAnnos(MBrecord, fixColNames=FALSE)
}
\arguments{
\item{MBrecord}{File name (and path) to a valid MassBank record file.}

\item{fixColNames}{Default \code{FALSE} uses the column names in the MassBank record,
i.e. \code{m/z}, \code{int.} and \code{rel.int.}. Use \code{TRUE} to make this more
R-friendly. The function \code{\link{fixMzIntColNames}} is used to do this. NOTE: if
\code{TRUE}, this is only performed after merging to ensure the merge works.}
}
\value{
Returns a data frame containing annotated peaks, or \code{NULL} if the retrieval failed.
}
\description{
This function extracts the PK$PEAK and PK$ANNOTATION entries from a MassBank record,
returning a merged data frame containing the corresponding entries. This uses the
\code{\link{getMBRecordPeakAnnotations}} and \code{\link{getMBRecordPeaks}} functions.
}
\details{
NOTE: for this to work properly, the m/z column names must match exactly.
This has only been verified so far on records generated with \code{\link{RMassBank}}.
}
\examples{
MBrecord <- system.file("extdata","EA020161_Diclofenac.txt",package="ReSOLUTION")
getMBPeaksAnnos(MBrecord)
getMBPeaksAnnos(MBrecord, fixColNames=TRUE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{getMBRecordPeakAnnotations}}, \code{\link{getMBRecordPeaks}},
\code{\link{fixMzIntColNames}}.
}

