% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassBankRecordeR.R
\name{getMBRecordEntry}
\alias{getMBRecordEntry}
\title{Get A Specific Entry from MassBank Record}
\usage{
getMBRecordEntry(field_code, record, trimEntry=TRUE)
}
\arguments{
\item{field_code}{The string to match to the MassBank record.}

\item{record}{The MassBank record in vector format to search, converted with
\code{\link{MBFileToVector}}.}

\item{trimEntry}{If \code{TRUE}, the \code{field_code} is removed, if \code{FALSE},
the full line is returned.}
}
\value{
Returns a vector containing the matching entries or \code{NULL} if the retrieval failed.
}
\description{
This function returns the entries in the MassBank record matching
the \code{field_code}, or \code{NULL} if absent. The record should be in vector
format, e.g. with \code{\link{MBFileToVector}}. Used to retrieve information in
\code{\link{getMBRecordInfo}}.
}
\details{
Use \code{\link{getMBRecordPeaks}}, \code{\link{getMBRecordPeakAnnotations}}
and/or \code{\link{getMBPeaksAnnos}} to retrieve peaks and peak annotations.
}
\examples{
MBrecord <- system.file("extdata","EA020161_Diclofenac.txt",package="ReSOLUTION")
record <- MBFileToVector(MBrecord)
field_code <- "CH$NAME"
field_code <- "CH$LINK: CAS"
field_code <- "CH$LINK CS"
field_code <- "CH$LINK: PUBCHEM CID"
getMBRecordEntry(field_code, record)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>.
}
\seealso{
\code{\link{MBFileToVector}}, \code{\link{getMBRecordInfo}}.
}

