% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFragConfigR.R
\name{getMBRecordInfo.MetFragConfig}
\alias{getMBRecordInfo.MetFragConfig}
\title{Extract Info from MassBank records for MetFrag Config files}
\usage{
getMBRecordInfo.MetFragConfig(MBrecord,peaklist_filepath,writePeaklist=TRUE)
}
\arguments{
\item{MBrecord}{Full path to the MassBank record containing peaks of interest}

\item{peaklist_filepath}{Full path to the peaklist that will be used in the MetFrag Config file. It is
recommeded to use a temporary file.}

\item{writePeaklist}{Default \code{TRUE} to extract and write peaklists to a temporary file. To use an existing
file, set to \code{FALSE}.}
}
\value{
Returns a list containing the peaklist filename, molecular formula, exact mass, precursor type,
ion mode and retention time
}
\description{
This is a convenience wrapper to extract info for MetFrag Config files if source mass spectra
are already in MassBank format.
}
\examples{

peaklist <- system.file("extdata","EA020161_Diclofenac_peaks.txt",package="ReSOLUTION")
MBrecord <- system.file("extdata","EA020161_Diclofenac.txt",package="ReSOLUTION")
MBinfo <- getMBRecordInfo.MetFragConfig(MBrecord,peaklist, writePeaklist=FALSE)
# use writePeaklist=TRUE to create a peaklist from the MassBank record

}
\seealso{
\code{\link{getMBRecordPeaks}}, \code{\link{getMBRecordEntry}}, \code{\link{MetFragConfig}}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
