% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassBankRecordeR.R
\name{getMBRecordInfo}
\alias{getMBRecordInfo}
\title{Summarize MassBank Entries in Directory of MassBank Records}
\usage{
getMBRecordInfo(directory, csv_name, field_codes="default", colNames="default",recursive=TRUE)
}
\arguments{
\item{directory}{The path to the directory to summarize. Subdirectories are also searched if
\code{recursive} is \code{TRUE}.}

\item{csv_name}{Name (with/without path) of the csv file to save the results.}

\item{field_codes}{A vector containing strings to match in the MassBank records. Setting
\code{"default"} will output default values, see \code{Details}.}

\item{colNames}{A vector containing column names to match \code{field_codes}. These must be
in the same order. If \code{"default"}, trimmed versions of the default \code{field_codes}
are used. If \code{"field_codes"}, then the entries in \code{field_codes} are used.}

\item{trimEntry}{If \code{TRUE}, the \code{field_code} is removed, if \code{FALSE},
the full line is returned.}
}
\value{
Saves the results in \code{csvfile}, or \code{error}.
}
\description{
This function returns a summary csv containing the entries in
all MassBank records in the given directory matching the list of \code{field_codes}.
Uses \code{\link{getMBRecordEntry}}.
}
\details{
Use \code{\link{getMBRecordPeaks}}, \code{\link{getMBRecordPeakAnnotations}}
and/or \code{\link{getMBPeaksAnnos}} to retrieve peaks and peak annotations. This function
uses \code{\link{getMBRecordEntry}} to retrieve individual entries from records. If \code{NULL},
then \code{NA} is printed to the csv. If multiple entries match (e.g. for \code{CH$NAME}), then
one entry in the csv is made, separated with a pipe (\code{|}).
This function will exit with an error if the length of \code{field_codes} and \code{colNames}
do not match.
The default settings are:
\code{field_codes <- c("ACCESSION", "CH$NAME", "CH$SMILES", "CH$EXACT_MASS", "CH$FORMULA", "CH$IUPAC",
"CH$LINK: CAS", "CH$LINK: PUBCHEM CID", "CH$LINK: INCHIKEY", "CH$LINK: CHEMSPIDER",
"AC$MASS_SPECTROMETRY: COLLISION_ENERGY", "AC$MASS_SPECTROMETRY: FRAGMENTATION_MODE",
"AC$CHROMATOGRAPHY: COLUMN_NAME", "AC$CHROMATOGRAPHY: RETENTION_TIME",
"MS$FOCUSED_ION: PRECURSOR_M/Z", "MS$FOCUSED_ION: PRECURSOR_TYPE")}
\code{colNames <- c("ACCESSION", "NAME", "SMILES", "NEUTRAL_EXACT_MASS", "FORMULA", "StdInChI",
"CAS_RN", "PUBCHEM_CID", "InChIKey", "CHEMSPIDER_ID", "COLLISION_ENERGY", "FRAGMENTATION_MODE",
"COLUMN_NAME", "RETENTION_TIME", "PRECURSOR_MZ", "PRECURSOR_TYPE")}
}
\examples{
directory <- system.file("extdata", "recdata", package = "ReSOLUTION")
# note this overwrites any file of the same name ...
csv_name <- paste0(getwd(),"/test.csv")
getMBRecordInfo(directory, csv_name, field_codes="default", colNames="default",recursive=TRUE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>, rewritten from getInfo
script from Erik Mueller.
}
\seealso{
\code{\link{MBFileToVector}}, \code{\link{getMBRecordEntry}}, \code{\link{getMBPeaksAnnos}}.
}

