% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassBankRecordeR.R
\name{getMBRecordPeakAnnotations}
\alias{getMBRecordPeakAnnotations}
\title{Retrieve Annotations from a MassBank Record}
\usage{
getMBRecordPeakAnnotations(MBrecord)
}
\arguments{
\item{MBrecord}{File name (and path) to a valid MassBank record file.}
}
\value{
Returns a data frame containing annotations, or \code{NULL} if the retrieval failed.
}
\description{
This function extracts the PK$ANNOTATION entries from a MassBank record,
returning a data frame containing the corresponding entries, with column headers taken
from the record. Use the function
\code{\link{getMBPeaksAnnos}} to get merged peak lists and annotations.
}
\examples{
MBrecord <- system.file("extdata","EA020161_Diclofenac.txt",package="ReSOLUTION")
getMBRecordPeakAnnotations(MBrecord)

}
\seealso{
\code{\link{getMBPeaksAnnos}}, \code{\link{getMBRecordPeaks}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>, based on code from Erik Mueller.
}
