% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassBankRecordeR.R
\name{getMBRecordPeaks}
\alias{getMBRecordPeaks}
\title{Get A Peak List from MassBank Record}
\usage{
getMBRecordPeaks(MBrecord, fixColNames = FALSE)
}
\arguments{
\item{MBrecord}{File name (and path) to a valid MassBank record file.}

\item{fixColNames}{Default \code{FALSE} uses the column names in the MassBank record,
i.e. \code{m/z}, \code{int.} and \code{rel.int.}. Use \code{TRUE} to make this more
R-friendly. The function \code{\link{fixMzIntColNames}} is used to do this.}
}
\value{
Returns a data frame containing peak data, or \code{NULL} if the retrieval failed.
}
\description{
This function extracts the peak list from a MassBank record, returning
a variable containing the peak m/z, intensity and relative intensity entries. The
column names are either original, or adjusted to be more R-friendly. Use the function
\code{\link{getMBPeaksAnnos}} to get peak lists with annotations.
}
\examples{
MBrecord <- system.file("extdata","EA020161_Diclofenac.txt",package="ReSOLUTION")
getMBRecordPeaks(MBrecord, fixColNames = FALSE)
getMBRecordPeaks(MBrecord, fixColNames = TRUE)

}
\seealso{
\code{\link{getMBPeaksAnnos}}, \code{\link{fixMzIntColNames}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>, based on code from Erik Mueller.
}
