% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_functions.R
\name{IDfromFilePath}
\alias{IDfromFilePath}
\title{Extract Compound ID from File Path and File Name}
\usage{
IDfromFilePath(file_path)
}
\arguments{
\item{file_path}{The file path to extract the ID from.}
}
\value{
Returns the compound ID or whatever content is in front of the first \code{"_"}.
Designed to work off output of other functions.
}
\description{
This extracts the Compound ID from the start of a
file name, given the full path.
}
\examples{
IDfromFilePath("C:/MY_DATA/MS1/3529_Indoxacarb_MS1raw.txt")

}
\seealso{
\code{\link{extractMS1.RMBarchive}}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
