% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassBankRecordeR.R
\name{MBFileToVector}
\alias{MBFileToVector}
\title{Read a MassBank Record (or any file) into a Vector}
\usage{
MBFileToVector(MBrecord)
}
\arguments{
\item{MBrecord}{File name (and path) to parse.}
}
\value{
Returns a vector containing the lines in the record
}
\description{
This opens the MassBank record (text file) in the
path \code{MBrecord} and uses \code{readLines} to save the values
in a vector
}
\examples{
MBrecord <- system.file("extdata","EA020161_Diclofenac.txt",package="ReSOLUTION")
MBFileToVector(MBrecord)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>, based on code from Erik Mueller.
}
