% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFragConfigR.R
\name{MetFragConfig.CompToxExportSDF}
\alias{MetFragConfig.CompToxExportSDF}
\title{Create MetFrag Config Files with LocalSDF and Scoring Terms from CompTox MetFrag SDF Export}
\usage{
MetFragConfig.CompToxExportSDF(mass, adduct_type, results_filename, peaklist_path, base_dir,
CompToxSDFtoLocalSDFterms, ...)
}
\arguments{
\item{mass}{The mass with which to search the candidate database (\code{DB}). Use \code{neutralPrecursorMass} and
\code{adduct_type} to set whether this is monoisotopic mass or an adduct species. Defaults to \code{adduct_type}.}

\item{adduct_type}{The adduct species used to define mass (if \code{neutralPrecursorMass=FALSE}) and fragmentation settings
in the config file, entered as either \code{PrecursorIonType} (text) or \code{PrecursorIonmode} (a number). The available
options are given in the system file \code{MetFragAdductTypes.csv} in the \code{extdata} folder.
Recommended default values (if ion state is unclear) are \code{[M+H]+} (1) for positive and \code{[M-H]-} (-1) for negative mode.}

\item{results_filename}{Enter a base filename for naming the results files - do not include file endings}

\item{peaklist_path}{Enter the full path and file name to the peak list for this config file}

\item{base_dir}{Enter the directory name to set up the subfolders for MetFrag batch results. If the folders don't exist,
subfolders \code{config}, \code{log} and \code{results} are created; the output of this function is saved in \code{config}.}

\item{CompToxSDFtoLocalSDFterms}{The output of \code{\link{CompToxSDFtoLocalSDFterms}}, used to set
\code{DB, localDB_path, UDS_Category and UDS_Weights} in \code{\link{MetFragConfig}}}
}
\value{
Returns a MetFrag config file name
}
\description{
This is a CompTox SDF Export-specific wrapper function for \code{\link{MetFragConfig}}.
}
\details{
Remaining parameters are described in \code{\link{MetFragConfig}}
}
\examples{
# Example from DOI: 10.1021/acs.est.7b01908
CompToxSDF <- system.file("extdata","CompToxBatchSearch_MetFrag_MSready_C10H14N2.sdf",package="ReSOLUTION")
LocalSDFterms <- CompToxSDFtoLocalSDFterms(CompToxSDF)
peaklist <- system.file("extdata","EQ300804_Nicotine_peaks.txt",package="ReSOLUTION")
test_dir <- "C:/DATA/Workflow/MetFrag22/metfrag_test_results"
config_file <- MetFragConfig.CompToxExportSDF(163.1230, "[M+H]+","Nicotine_PrecMass_MpHp_SDF",peaklist, test_dir, LocalSDFterms)
metfrag_dir <- "C:/DATA/Workflow/MetFrag22/"
MetFragCL_name <- "MetFrag2.4.4-msready-CL.jar"
runMetFrag(config_file, metfrag_dir, MetFragCL_name)

}
\seealso{
\code{\link{MetFragConfig}}, \code{\link{runMetFrag}}, \code{\link{CompToxSDFtoLocalSDFterms}}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu> in partnership with Christoph Ruttkies (MetFragCL),
Antony J. Williams and team (CompTox Dashboard)
}
