% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{NewNodesGroup}
\alias{NewNodesGroup}
\title{NewNodesGroup}
\usage{
NewNodesGroup(OldCom, NewCom, NodesGroups)
}
\arguments{
\item{OldCom}{the community which the node have been.}

\item{NewCom}{the community which the node moved to.}

\item{NodesGroups}{table stores the number of the community to which a particular node is currently assigned.}
}
\description{
Updates table NodesGroups.
}

