% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nontargetWrap.R
\name{nontargetWrap}
\alias{nontargetWrap}
\title{Wrapper Function for Package nontarget}
\usage{
nontargetWrap(peaklist_file, desc, path, pos=TRUE, inclHomol=FALSE,
colnums=c(1,3,5))
}
\arguments{
\item{peaklist_file}{File containing the peak list}

\item{desc}{Description - used to create file names}

\item{path}{Path in which to save the results files}

\item{pos}{Define the mode. Default \code{TRUE} indicates positive data,
\code{FALSE} indicates negative data.}

\item{inclHomol}{Indicate whether to include homologue series detection. Default \code{FALSE}.
If \code{TRUE}, all plus certain selected series are screened. Can be very
time consuming and lead to memory errors.}

\item{colnums}{Define the location of the mz, int and RT columns in \code{peaklist_file}.
The default \code{c(1,3,5)} works on enviPick output. Must be redefined for other peaklists
to avoid strange results!}
}
\value{
Results are saved into many output files in \code{path}
}
\description{
This wrapper function runs the nontarget workflow with
selected default settings for positive and negative mode, creating
many output files to interpret the results.
}
\examples{

#Note: this produces a lot of files, caution before trying.
peaklist_path <- system.file("extdata","Blind_nano_IS_toPeak750.txt",package="ReSOLUTION")
test_path <- paste0(getwd(),"/nontargettest")
dir.create(test_path)
nontargetWrap(peaklist_path, "BlindNanoIS_testWrap_to750", test_path, pos=TRUE, inclHomol=FALSE, colnums=c(1,3,5))
gc() #clear memory after.


}
\author{
Emma Schymanski (<emma.schymanski@uni.lu>, wrapper), Martin Loos
(\code{\link{nontarget}}).
}
\seealso{
\code{\link{nontarget}}.
}

