% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EICandMSMS.R
\name{plotAdductEICs}
\alias{plotAdductEICs}
\title{Plot EICs from Adduct List}
\usage{
plotAdductEICs(mzML_filePath, smiles, adduct_list, rt=0, rt_in_sec=TRUE, plot_title=NULL,
rt_window=NULL, mz_limit=0.001, isPos=TRUE, isLog=FALSE, isSwitching=FALSE, greyscale=FALSE,
colours=NULL, cpdID=NULL, ymin_log=3, switchShift=0, kekulise=TRUE)
}
\arguments{
\item{mzML_filePath}{Path to the mzML (or mzXML) file to extract the EICs.}

\item{smiles}{SMILES code of the compound of interest.}

\item{adduct_list}{Vector of adducts to process, matching entries in the \code{adducts}
table in \code{\link{enviPat}}. Used to fill \code{mz} and \code{names} in \code{plotEICs}.}

\item{rt}{The retention time (in seconds) to centre EIC extraction. If unknown, set
\code{rt_window} to \code{NULL}.}

\item{rt_in_sec}{Default \code{TRUE} indicates that \code{rt} is in seconds. If
\code{FALSE}, \code{rt} is converted to seconds internally.}

\item{plot_title}{Title for the plot. If \code{NULL}, \code{basename(mzML_filePath)} is used.}

\item{rt_window}{The retention time window (in seconds) for EIC extraction. This is added
to \code{rt} to form the extraction window. Should not be zero! If \code{NULL},
the whole EIC is extracted.}

\item{isPos}{\code{TRUE} indicates positive mode; \code{FALSE} negative mode. Only
relevant for switching data.}

\item{isLog}{If \code{TRUE}, EICs are plotted with a log scale; \code{FALSE} with
absolute intensities. Zero intensities are lost with \code{TRUE}. Acting strangely.}

\item{isSwitching}{If \code{TRUE}, will split switching files assuming that the
first scan is positive and that positive and negative alternate, unless \code{switchShift}
is set. USE WITH CAUTION.}

\item{greyscale}{If \code{TRUE}, greyscale plot will be produced. If \code{FALSE},
rainbow colours are assigned automatically.}

\item{colours}{Used to define EIC colours. If \code{NULL}, rainbow or greyscale colours are used,
depending on \code{greyscale}.}

\item{cpdID}{If given, this is used in the output. Otherwise a value of 1 is assigned.}

\item{ymin_log}{This defines the y axis intercept for log plots, for aesthetic reasons. Should be
just under the \code{log(noise_intensity)} threshold. \code{3} is appropriate for Orbitraps.}

\item{switchShift}{If \code{isSwitching} is \code{TRUE}, this can be used to shift the start scan between
positive and negative mode. A \code{"quick fix"}, USE WITH CAUTION.}

\item{kekulise}{Controls aromaticity detection in SMILES for plotting.}

\item{addMSMS}{Indicates whether MSMS scans should be added to plot with default settings \code{TRUE}.
\code{FALSE} does not attempt to add MSMS scans.}

\item{newRMB}{Indicates whether new RMassBank version is used for MS/MS retrieval (\code{TRUE}).
Set to \code{FALSE} for older versions (< 2.0.0) or if error is observed.}

\item{plot_title}{If \code{NULL}, title is autogenerated using \code{isPos}, \code{smiles} and
\code{cpdID} if given.}
}
\value{
Returns a plot in the active device and a summary output.
}
\description{
This plots EICs of adducts in \code{adduct_list} from the mzML (or mzXML)
and \code{smiles} given, with various options. Adduct masses are calculated from the
\code{smiles} using \code{\link{getSuspectMasses}} via \code{\link{rcdk}}. Labelling
will not be processed correctly. Adducts that cannot be calculated (especially in negative mode)
are automatically skipped. Uses \code{\link{plotEICs}} behind the scenes.
}
\examples{
pos_add <- c("M+H", "M+", "M+Na", "M+K", "M+NH4")
neg_add <- c("M-H", "M-", "M+FA-H", "M+Cl", "M-H2O-H")
library(RMassBank)
library(RMassBankData)
RmbDefaultSettings()
mzML_files <- list.files(system.file("spectra", package="RMassBankData"), ".mzML", full.names = TRUE)
mzML_file <- mzML_files[1]
cmpd_id <- "2818"
mzML_file <- mzML_files[5]
cmpd_id <- "2822"
loadList(system.file("list/NarcoticsDataset.csv",package="RMassBankData"))
smiles <- findSmiles(cmpd_id)
rt <- as.numeric(findRt(cmpd_id))
plotAdductEICs(mzML_file, pos_add, rt, rt_in_sec=FALSE, smiles = smiles)
plotAdductEICs(mzML_file, pos_add, rt, rt_in_sec=FALSE, smiles = smiles, addMSMS=TRUE)
# at this stage, log plots are not working properly...
plotAdductEICs(mzML_file, pos_add, rt, rt_in_sec=FALSE, smiles = smiles, isLog=TRUE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>, Michael Stravs, Benedikt Lauper
}
\seealso{
\code{\link{getSuspectMasses}}, \code{\link{findEIC}}, \code{\link{plotEICs}}.
}

