% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EICandMSMS.R
\name{plotEICs}
\alias{plotEICs}
\title{Plot EICs from mz and RT List}
\usage{
plotEICs(mzML_filePath, mz, rt, rt_in_sec=TRUE, names=NULL, smiles="", plot_title=NULL,
rt_window=NULL, mz_limit=0.001, ppm=10, isPos=TRUE, isLog=FALSE, isSwitching=FALSE,
greyscale=FALSE, colours = NULL, cpdID=NULL, ymin_log=3, switchShift=0,
kekulise=TRUE, addMSMS=TRUE, newRMB=TRUE, takePPM=TRUE)
}
\arguments{
\item{mzML_filePath}{Path to the mzML (or mzXML) file to extract the EICs.}

\item{mz}{Vector of mz values to process.}

\item{rt}{Vector of retention times to process. Set the unit with \code{rt_in_sec}.}

\item{rt_in_sec}{Default \code{TRUE} indicates that \code{rt} is in seconds. If
\code{FALSE}, \code{rt} is converted to seconds internally.}

\item{names}{Optional names to belong to each \code{mz}. If \code{NULL}, \code{mz} is used.}

\item{smiles}{SMILES code of the compound of interest.}

\item{plot_title}{Title for the plot. If \code{NULL}, \code{basename(mzML_filePath)} is used.}

\item{rt_window}{The retention time window (in seconds) for EIC extraction. This is added
to \code{rt} to form the extraction window. Should not be zero! If \code{NULL},
the whole EIC is extracted.}

\item{mz_limit}{Default 0.001. This controls the limits for EIC extraction. Choose between
using \code{mz_limit} and \code{ppm} with \code{takePPM}.}

\item{ppm}{Default 10 ppm is an alternative to \code{mz_limit} for EIC extraction.}

\item{isPos}{\code{TRUE} indicates positive mode; \code{FALSE} negative mode. Only
relevant for switching data.}

\item{isLog}{If \code{TRUE}, EICs are plotted with a log scale; \code{FALSE} with
absolute intensities. Zero intensities are lost with \code{TRUE}.}

\item{isSwitching}{If \code{TRUE}, will split switching files assuming that the
first scan is positive and that positive and negative alternate, unless \code{switchShift}
is set. USE WITH CAUTION.}

\item{greyscale}{If \code{TRUE}, greyscale plot will be produced. If \code{FALSE},
rainbow colours are assigned automatically.}

\item{colours}{Used to define EIC colours. If \code{NULL}, rainbow or greyscale colours are used,
depending on \code{greyscale}.}

\item{cpdID}{If given, this is used in the output. Otherwise a value of 1 is assigned.}

\item{ymin_log}{This defines the y axis intercept for log plots, for aesthetic reasons. Should be
just under the \code{log(noise_intensity)} threshold. \code{3} is appropriate for Orbitraps.}

\item{switchShift}{If \code{isSwitching} is \code{TRUE}, this can be used to shift the start scan between
positive and negative mode. A \code{"quick fix"}, USE WITH CAUTION.}

\item{kekulise}{Controls aromaticity detection in SMILES for plotting.}

\item{addMSMS}{Indicates whether MSMS scans should be added to plot with default settings \code{TRUE}.
\code{FALSE} does not attempt to add MSMS scans.}

\item{newRMB}{Indicates whether new RMassBank version is used for MS/MS retrieval (\code{TRUE}).
Set to \code{FALSE} for older versions (< 2.0.0) or if error is observed.}

\item{takePPM}{Indicates whether ppm or mz_limit value should be used to extract EICs.}

\item{plot_title}{If \code{NULL}, title is autogenerated using \code{isPos}, \code{smiles} and
\code{cpdID} if given.}
}
\value{
Returns a plot in the active device and a summary output.
}
\description{
This plots EICs of given \code{mz} and from the mzML (or mzXML),
with various options. Additional wrapper functions provide different ways of generating mz.
}
\examples{
library(RMassBank)
library(RMassBankData)
RmbDefaultSettings()
mzML_files <- list.files(system.file("spectra", package="RMassBankData"), ".mzML", full.names = TRUE)
mzML_file <- mzML_files[1]
cmpd_id <- "2818"
loadList(system.file("list/NarcoticsDataset.csv",package="RMassBankData"))
smiles <- findSmiles(cmpd_id)
mz <- as.numeric(findMz(cmpd_id, mode="pH")[3])
rt <- as.numeric(findRt(cmpd_id))
plotEICs(mzML_file, mz, rt, rt_in_sec=FALSE, smiles = smiles)
plotEICs(mzML_file, mz, rt, rt_in_sec=FALSE, smiles = smiles, addMSMS=TRUE)


}
\author{
Emma Schymanski <emma.schymanski@uni.lu>, Michael Stravs, Benedikt Lauper
}
\seealso{
\code{\link{findEIC}}.
}

