% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSMS.R
\name{plotMS}
\alias{plotMS}
\title{Plot MS1 Spectrum}
\usage{
plotMS(mz, int, main, labels=NULL, formatFormula=TRUE, absInt=FALSE,ylim_factor=2)
}
\arguments{
\item{mz}{Vector containing mz values (x) to plot.}

\item{int}{Vector containing intensity values (y) to plot.}

\item{main}{Title for the plot.}

\item{labels}{Vector containing labels for selected peaks. If \code{NULL}, no labels
are plotted.}

\item{formatFormula}{If \code{TRUE}, uses \code{chemistry2expression} to create subscript numbers.
Any \code{+} or \code{-} are removed to avoid errors. If \code{FALSE}, \code{labels} are printed as is.
This is passed to \code{chem2express} in \code{\link{plotSpectra}}.}

\item{absInt}{If \code{TRUE}, absolute intensity values are used. If \code{FALSE}, relative intensities
are calculated on individual spectra, scaled to 1000.}

\item{ylim_factor}{Scaling factor for y-axis. Default \code{2.5} leaves ample space for structure,
annotations and legend. Reduce for peaks only. Passed to \code{yfactor} in \code{\link{plotSpectra}}.}
}
\value{
Returns a plot in the current plotting device.
}
\description{
This is a wrapper function for \code{\link{plotSpectra}} for backwards
compatability with previous workflows/scripts. \code{\link{plotSpectra}} offers
much more functionality. \code{xlim} and \code{ylim} are fixed suited to MS1 spectra.
}
\examples{
ms1_mz <- c(210.1125, 211.1154, 212.1187)
ms1_int <- c(420889280, 52790856, 2265951.5)
ms1_labels <- c("[M+H]+", "13C-M+1", "13C-M+2")
smiles_1 <- "O=C(Oc1ccccc1OC(C)C)NC"
plotMS(ms1_mz, ms1_int, main="test MS1 spec",labels=ms1_labels, formatFormula=FALSE,
absInt=TRUE)
plotMS(ms1_mz, ms1_int, main="test MS1 spec",labels=ms1_labels, formatFormula=FALSE,
absInt=FALSE, smiles=smiles_1)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{plotSpectra}}, \code{\link{renderSMILES.rcdk}}, \code{\link{chemistry2expression}},
\code{\link{trimAnnotation}}.
}

