% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSMS.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Plot Multiple Spectra on a Single Plot}
\usage{
plotSpectra(mz, int, main, labels=NULL,labelSpec=1,label_mz=NA,
label_int=NA,mz_a=NA,int_a=NA, absInt=FALSE, logInt=FALSE, mz_b=NA,int_b=NA,
smiles="",kekulise=TRUE,legend=c("spec1","spec2","spec3"),
xlim=NULL,ylim=NULL, yfactor=2.5, line_colour=c("black","red","green"),
line_type=c("solid","dashed","dotdash"),line_width=c(2,2,1),
chem2express=FALSE, srt=60, pos=4,offset=-0.2)
}
\arguments{
\item{mz}{Vector containing mz values (x) to plot first.}

\item{int}{Vector containing intensity values (y) to plot first.}

\item{main}{Title for the plot.}

\item{labels}{Vector containing labels for selected peaks. If \code{NULL}, no labels
are plotted.}

\item{labelSpec}{Indicates which spectrum to use for label coordinates. Must be same
length as \code{labels} to avoid recycling labels. If \code{1}, \code{mz} and \code{int}
are used, if \code{2} then \code{mz_a} and \code{int_a}, if \code{3}, then \code{mz_b}
and \code{int_b}. Parameters \code{label_mz} and \code{label_int} take precedence.}

\item{label_mz}{Values defining starting x coordinates for \code{labels}.}

\item{label_int}{Values defining starting y coordinates for \code{labels}.}

\item{mz_a}{Vector containing mz values for second spectrum.}

\item{int_a}{Vector containing intensity values for second spectrum.}

\item{absInt}{If \code{TRUE}, absolute intensity values are used. Be careful that all
spectra match! If \code{FALSE}, relative intensities are calculated on individual
spectra, scaled to 1000.}

\item{logInt}{If \code{FALSE}, intensity values are used as is. If \code{TRUE}, log
values are used; this can be useful to find small peaks.}

\item{mz_b}{Vector containing mz values for third spectrum.}

\item{int_b}{Vector containing intensity values for third spectrum.}

\item{smiles}{SMILES code of the structure to plot. Leave empty for no structure.}

\item{kekulise}{Controls aromaticity detection of \code{smiles}. Default \code{TRUE}
should work best with rcdk>3.4.1. Try \code{FALSE} if rings and double bonds appear together
in aromatic rings.}

\item{legend}{Vector containing legend entries.}

\item{xlim}{Option to control x-axis limits. If \code{NULL}, \code{xlim=c(0,1.1*max(all_mz))}}

\item{ylim}{Option to control y-axis limits. If \code{NULL}, \code{ylim=c(0,yfactor*max(all_int))}
unless log scale is used, where \code{ylim=c(-1, yfactor+8)}.}

\item{yfactor}{Scaling factor for y-axis. Default \code{2.5} leaves ample space for structure,
annotations and legend. Reduce for peaks only.}

\item{line_colour}{Optional vector to re-define default line colours.}

\item{line_type}{Optional vector to re-define default line type.}

\item{line_width}{Optional vector to re-define default line widths.}

\item{chem2express}{If \code{TRUE}, uses \code{chemistry2expression} to create subscript numbers.
Any \code{+} or \code{-} are removed to avoid errors. If \code{FALSE}, \code{labels} are printed as is.}

\item{srt}{Optional parameter to adjust \code{srt} value for \code{labels}.}

\item{pos}{Optional parameter to adjust \code{pos} value for \code{labels}.}

\item{offset}{Optional parameter to adjust \code{offset} value for \code{labels}.}
}
\value{
Returns a plot in the current plotting device.
}
\description{
This is the main function to plot multiple mass spectra
on one plot, with many options to control the output.
}
\details{
For controlling label orientation, the following settings are useful:
Default (angle): \code{srt=60}, \code{pos=4}, \code{offset=-0.2};
Vertical above peak: \code{srt=90}, \code{pos=4}, \code{offset=0};
Horizontal, centred above peak: \code{srt=0}, \code{pos=3}, \code{offset=0.2}.

To avoid overlapping labels, use \code{\link{trimAnnotation}}.
}
\examples{
mz_1 <- c(58.0287, 111.0441, 168.0655, 210.1125)
int_1 <- c(23.0000, 999.0000, 843.5855, 999.0000)
labels_1 <- c("C2H4NO", "C6H7O2", "C8H10NO3", "C11H16NO3")
smiles_1 <- "O=C(Oc1ccccc1OC(C)C)NC"
plotSpectra(mz_1, int_1, main="test spec",labels=labels_1, labelSpec=1, smiles=smiles_1)
plotSpectra(mz_1, int_1, main="test spec",labels=labels_1, labelSpec=1, smiles=smiles_1,
chem2express=TRUE, mz_a=mz_1, int_a = int_1, legend=c("spec1","spec1 again"))

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{plot}}, \code{\link{renderSMILES.rcdk}}, \code{\link{chemistry2expression}},
\code{\link{text}}, \code{\link{trimAnnotation}}.
}

