% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSDF.R
\name{process.sdf.file}
\alias{process.sdf.file}
\title{Process SDF File to retrieve tags and properties}
\usage{
process.sdf.file(SDF_file)
}
\arguments{
\item{SDF_file}{Full path and file name to SDF file}
}
\value{
This returns a set of SDF tags and information whether values are numeric or not
}
\description{
This parses an SDF file and retrieves information used to select relevant to MetFrag
scoring terms from CompTox Dashboard Export files. Used within \code{\link{CompToxSDFtoLocalSDFterms}}.
}
\examples{
CompToxSDF <- system.file("extdata","CompToxBatchSearch_MetFrag_MSready_C10H14N2.sdf",package="ReSOLUTION")
process.sdf.file(CompToxSDF)

}
\seealso{
\code{\link{CompToxSDFtoLocalSDFterms}}
}
\author{
Christoph Ruttkies, Emma Schymanski <emma.schymanski@uni.lu>
}
