% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMB_EIC_prescreen.R
\name{RMB_EIC_prescreen}
\alias{RMB_EIC_prescreen}
\title{EIC Prescreening function for RMassBank}
\usage{
RMB_EIC_prescreen(archive_name, RMB_mode, FileList, cmpd_list,
  ppm_limit_fine = 10, EIC_limit = 0.001)
}
\arguments{
\item{archive_name}{Name of the archive (compatible with RMassBank naming)}

\item{RMB_mode}{Mode in RMassBank}

\item{FileList}{The FileList to be used in RMassBank}

\item{cmpd_list}{The CompoundList to be used in RMassBank}

\item{ppm_limit_fine}{The fine limit for extracting the MS/MS for \code{findMsMsHR.mass} in 
RMassBank (ppm)}

\item{EIC_limit}{The (absolute) limit for EIC extraction.}
}
\value{
Function returns a PDF and CSV summary file.
}
\description{
This is a wrapper function to perform prescreening of
EICs and MS/MS scans prior to running RMassBank, so entries can 
be checked visually. Output is a PDF and summary CSV; this is designed
to check the data and not look pretty. 
This is mainly an internal function made public just in case it helps anyone. 
It is recommended to run this immediately prior to msmsWorkflow in RMassBank
once all the appropriate files are set up.
}
\examples{

Currently no examples implemented.

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}

