% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuspectScreening.R
\name{screen.homol}
\alias{screen.homol}
\title{Screen EICs for Homologue Series}
\usage{
screen.homol(file, file_desc, series, series_name, adduct, rt_window=NULL, ppm=10,
greyscale=FALSE, smiles="", formula=NULL, addMSMS=FALSE, isLog=FALSE,
isPos=TRUE, kekulise=TRUE, writePDF=FALSE, writeCSV=FALSE)
}
\arguments{
\item{file}{Path and filename to mz(X)ML file to screen}

\item{file_desc}{Description for the file (used in plot title and file names)}

\item{series}{The series from \code{\link{build.homol}} to screen}

\item{series_name}{The name of the series (used in legend)}

\item{adduct}{The adduct species of the homologue to screen. NOTE: the adduct must be exactly as written 
in the adducts table in the enviPat package.}

\item{rt_window}{The retention time window (in seconds) for EIC extraction. This is added
to \code{rt} to form the extraction window. Should not be zero! If \code{NULL},
the whole EIC is extracted.}

\item{ppm}{The ppm limit for EIC extraction}

\item{greyscale}{If \code{TRUE}, greyscale plot will be produced. If \code{FALSE},
rainbow colours are assigned automatically.}

\item{smiles}{SMILES code of the compound of interest. Can also be generic SMILES
(this needs to be tested...).}

\item{formula}{A text field to put the formula of the series if no SMILES.}

\item{addMSMS}{Indicates whether MSMS scans should be added to plot with default settings \code{TRUE}.
\code{FALSE} does not attempt to add MSMS scans.}

\item{isLog}{If \code{TRUE}, EICs are plotted with a log scale; \code{FALSE} with
absolute intensities. Zero intensities are lost with \code{TRUE}. TRUE may not work properly.}

\item{isPos}{Setting to pass on to \code{\link{plotEICs}}. \code{TRUE} indicates positive mode;
\code{FALSE} negative mode. Only relevant for switching data.}

\item{kekulise}{Controls aromaticity detection in SMILES for plotting.}

\item{writePDF}{Writes the plot to PDF rather than active device. PDF name is autocreated.}

\item{writeCSV}{Writes the screen summary to CSV - file name is autocreated.}
}
\value{
Returns a plot in the active device, or PDF and summary output.
}
\description{
This is a wrapper for plotEICs designed for screening homologue series
built from \code{\link{build.homol}}.
}
\examples{
pegs <- build.homol("C4H10O3","C2H4O","PEGS",2,14)
library(RMassBank)
library(RMassBankData)
RmbDefaultSettings()
mzML_files <- list.files(system.file("spectra", package="RMassBankData"), ".mzML", full.names = TRUE)
mzML_file <- mzML_files[1]
screen.homol(mzML_file, "test_file", pegs, "PEGS_MpH", "M+H")
screen.homol(mzML_file, "test_file", pegs, "PEGS_MpNH4", "M+NH4")
#Note this doesn't look pretty as PEGS aren't present, but does test that it works.


}
\seealso{
\code{\link{plotEICs}}, \code{\link{build.homol}}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
