% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuspectScreening.R
\name{screenSuspects}
\alias{screenSuspects}
\title{Screen Suspects in Peak List by Mass}
\usage{
screenSuspects(peaklist, suspects,ppmMax=5)
}
\arguments{
\item{peaklist}{A peak list containing columns \code{mz}, \code{RT}, \code{Int}.}

\item{suspects}{A list of suspects containing columns \code{Name} and \code{Suspect_mass}.}

\item{ppmMax}{Maximum \code{ppm} to match masses.}
}
\value{
Returns an ordered list of suspect hits.
}
\description{
This screens a peak list for given suspects by exact mass
}
\examples{

}
\author{
Jennifer Schollee, Michael Stravs, Emma Schymanski <emma.schymanski@uni.lu>
}

