% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSMS.R
\name{SpecComp.OrgMassSpecR}
\alias{SpecComp.OrgMassSpecR}
\title{Compare Mass Spectra (basic wrapper for OrgMassSpecR function)}
\usage{
SpecComp.OrgMassSpecR(spec1, spec2, output_file_name, t=0.005, b=0.1, header=FALSE)
}
\arguments{
\item{spec1}{File containing \code{mz} and \code{intensity} values of first (top) spectrum.}

\item{spec2}{File containing \code{mz} and \code{intensity} values of second (bottom) spectrum.}

\item{output_file_name}{File name for the output table. If empty no output is written.}

\item{t}{Numeric value specifying the tolerance used to align the m/z values of the two spectra.}

\item{b}{Numeric value specifying the baseline threshold for peak identification.
Expressed as a percent of the maximum intensity.}

\item{header}{Sets whether \code{spec1} and \code{spec2} contain header row (\code{TRUE} or not (\code{FALSE}).}
}
\value{
Returns a similarity value, a file containing the comparison table and a plot in the active device.
}
\description{
This uses the \code{\link{SpectrumSimilarity}} function of
\code{OrgMassSpecR} to calculate a similarity value, output the table to
a separate text file and a plot to the active device
}
\examples{
spec1 <- system.file("extdata","EA000405_peaks.txt",package="ReSOLUTION")
spec2 <- system.file("extdata","EA000407_peaks.txt",package="ReSOLUTION")
sim <- SpecComp.OrgMassSpecR(spec1,spec2, header=F)
sim

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>,
Nathan Dodder (\code{\link{SpectrumSimilarity}} function).
}
\seealso{
\code{\link{SpectrumSimilarity}}.
}

