% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_functions.R
\name{trimAnnotation}
\alias{trimAnnotation}
\title{Select a Subset of Annotated Peaks for Plotting}
\usage{
trimAnnotation(annotated_dataframe,mzLimit=10,IntLimit=1,
column_names = c("mz","anno","intensity","relInt"), UseRmbDefaultSettings=TRUE)
}
\arguments{
\item{annotated_dataframe}{The data frame containing the annotated data to trim}

\item{mzLimit}{The \code{mz} window to exclude other annotations.}

\item{IntLimit}{The default value of 1 is recommended. See details.}

\item{column_names}{This should be used to defined the column names and can be retrieved from
\code{annotated_dataframe} using \code{\link{colnames}}. NOTE that the columns \code{mz} and
\code{intensity} MUST exist for this function to work.}

\item{UseRmbDefaultSettings}{The default \code{TRUE} calls \code{RmbDefaultSettings}; if
other settings are in use set this to \code{FALSE}. See Details.}
}
\value{
Returns a "trimmed" annotated_dataframe
}
\description{
This uses the \code{\link{filterPeakSatellites}} of \code{\link{RMassBank}}
to select annotated peaks for plotting and thus prevent overlap.
}
\details{
\code{UseRmbDefaultSettings}: If \code{RMassBank} is not in use for any other purpose, the
default \code{TRUE} calls \code{RmbDefaultSettings} to ensure that settings are present for use.
If other settings have been defined using \code{RMassBank}, set this to \code{FALSE} to avoid
overwriting those settings. This function retrieves settings and adjusts them internally without
overwriting any stored settings.
\code{IntLimit}: This is used in \code{\link{filterPeakSatellites}} and is relative to 1.
For this context, this should be kept at 1. Values below 1 will let additional peaks through;
values above 1 lead to strange behaviour.
}
\examples{
MBrecord <- system.file("extdata","EA000405.txt",package="ReSOLUTION")
MBrecord <- system.file("extdata","EA026206_Simazine.txt",package="ReSOLUTION")
anno_peaks <- getMBPeaksAnnos(MBrecord, fixColNames=TRUE)
anno_cols <- sub("Int","intensity",colnames(anno_peaks))
colnames(anno_peaks) <- anno_cols
trimAnnotation(anno_peaks, mzLimit=10, IntLimit=1, column_names=anno_cols)


}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{extractMS1.RMBarchive}}
}

