% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_resolution.R
\name{cluster_resolution}
\alias{cluster_resolution}
\title{cluster_resolution}
\usage{
cluster_resolution(graph, t = 1, directed = FALSE)
}
\arguments{
\item{graph}{An igraph network or a data frame of three columns: source, target, and weights.}

\item{t}{The time-scale parameter of the process which uncovers community structures at different resolutions.}

\item{directed}{Logical. TRUE if the network is directed. Ignored if graph is an igraph object.}
}
\value{
Table with information about community which has been found for each node.
}
\description{
cluster_resolution function has been created based on paper "Laplacian dynamics and Multiscale Modular Structure in Networks" R. Lambiotte et al.
               Algorithm which finds communities using stability as an objective function to be optimised
                 in order to find the best partition  of network.
}
\examples{
library(igraph)
g <- nexus.get("miserables")
cluster_resolution(g,directed=FALSE,t=1)
}

